pyproject.toml
[project]
name = "litellm-proxy-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "gitpython>=3.1.46",
    "litellm[proxy]>=1.81.11",
    "watchdog>=6.0.0",
]
